adk_pipeline/requirements.txt
python-dotenv
google-adk
requests
litellm